Modularización del view Detail,conexion con back mediante action y modif geid back
