+++ client/src/redux/actions/index.js
@@ -18,7 +18,7 @@ export const changeTheme=(payload)=>{
 export function getdetailid(id) {
     return async function (dispatch) {
       try {
-        var json = await axios.get(`${id}`);
+        var json = await axios.get(`http://localhost:3001/${id}`);
         return dispatch({
           type: types.GET_DETAILS,
           payload: json.data,
